@@ -1,3 +1,3 @@
 module github.com/joschahenningsen/tum-live/worker/edge
 
-go 1.18
+go 1.19
